backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.69",
    "fastapi>=0.121.1",
    "langchain-openai>=1.1.7",
    "langchain-perplexity>=1.1.0",
    "mangum>=0.19.0",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
